+++ PowerPBS/Lib/PowerPBSShadowCasterPass.hlsl
@@ -9,19 +9,19 @@
 #include "PowerPBSInput.hlsl"
 #include "PowerPBSUrpShadows.hlsl"
 
-float3 _LightDirection;
+half3 _LightDirection;
 
 struct v2f{
-    float2 uv:TEXCOORD0;
-    float4 pos:SV_POSITION;
+    half2 uv:TEXCOORD0;
+    half4 pos:SV_POSITION;
 };
 
 //--------- shadow helpers
 #if defined(URP_SHADOW)
-float4 GetShadowPositionHClip(appdata_full input){
-    float3 worldPos = mul(unity_ObjectToWorld,input.vertex);
-    float3 worldNormal = UnityObjectToWorldNormal(input.normal);
-    float4 positionCS = UnityWorldToClipPos(ApplyShadowBias(worldPos,worldNormal,_LightDirection));
+half4 GetShadowPositionHClip(appdata_full input){
+    half3 worldPos = mul(unity_ObjectToWorld,input.vertex);
+    half3 worldNormal = UnityObjectToWorldNormal(input.normal);
+    half4 positionCS = UnityWorldToClipPos(ApplyShadowBias(worldPos,worldNormal,_LightDirection));
     #if UNITY_REVERSED_Z
         positionCS.z = min(positionCS.z, positionCS.w * UNITY_NEAR_CLIP_VALUE);
     #else
@@ -44,9 +44,9 @@ v2f vert(appdata_full input){
     return output;
 }
 
-float4 frag(v2f input):SV_Target{
+half4 frag(v2f input):SV_Target{
     if(_AlphaTestOn){
-        float4 tex = SAMPLE_TEXTURE2D(_MainTex,sampler_MainTex,input.uv);
+        half4 tex = SAMPLE_TEXTURE2D(_MainTex,sampler_MainTex,input.uv);
         clip(tex.a - _Cutoff);
     }
     return 0;
