+++ PowerPBS/Testcase/pbr1/Lib/PBRForwardPass.hlsl
@@ -7,15 +7,15 @@
 #include "Lib/PBRInput.hlsl"
 #include "Lib/Core/Fog.hlsl"
 
-float3 CalcIBL(float3 viewDir, float3 n,float a){
+half3 CalcIBL(half3 viewDir, half3 n,half a){
     a = a* (1.7 - a * 0.7);
-    float mip = round(a * 6);
-    float3 reflectDir = reflect(-viewDir,n);
-    float4 hdrEnv = SAMPLE_TEXTURECUBE_LOD(unity_SpecCube0,samplerunity_SpecCube0,reflectDir,mip);
+    half mip = round(a * 6);
+    half3 reflectDir = reflect(-viewDir,n);
+    half4 hdrEnv = SAMPLE_TEXTURECUBE_LOD(unity_SpecCube0,samplerunity_SpecCube0,reflectDir,mip);
     return DecodeHDR(hdrEnv,unity_SpecCube0_HDR);
 }
 
-float3 CalcGI(){
+half3 CalcGI(){
     return 0;
 }
 #endif //PBR_FORWARD_PASS_HLSL
